backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.0
pymysql>=1.1.0
alembic>=1.12.1
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.25.2
bcrypt>=4.1.2
email-validator>=2.0.0
